♻️ refactor: move PhysicalMetric entity to athlete domain
- Move physical-metric.entity.ts from athletes/physical-metric/ to athletes/athlete/domain/
- Update import paths in athlete.entity.ts and physical-metric.seeder.ts
- Consolidate physical metric functionality within athlete domain
