+++ apps/api/src/modules/athletes/athlete/domain/physical-metric.entity.ts
@@ -1,5 +1,5 @@
 import { Entity, ManyToOne, PrimaryKey, Property } from '@mikro-orm/core';
-import { Athlete } from '../athlete/domain/athlete.entity';
+import { Athlete } from './athlete.entity';
 
 @Entity()
 export class PhysicalMetric {
